Moved all left components to core
