+++ src/main/kotlin/dev/syoritohatsuki/duckyupdaterrework/client/message/UpdateAvailable.kt
@@ -2,7 +2,7 @@ package dev.syoritohatsuki.duckyupdaterrework.client.message
 
 import dev.syoritohatsuki.duckyupdaterrework.core.dto.durw.Printer
 import dev.syoritohatsuki.duckyupdaterrework.core.dto.modrinth.AdditionalInfo
-import dev.syoritohatsuki.duckyupdaterrework.storage.ModId
+import dev.syoritohatsuki.duckyupdaterrework.core.storage.ModId
 import net.minecraft.text.ClickEvent
 import net.minecraft.text.HoverEvent
 import net.minecraft.text.Text
