+++ src/main/kotlin/dev/syoritohatsuki/duckyupdaterrework/core/DuckyUpdaterApi.kt
@@ -2,9 +2,9 @@ package dev.syoritohatsuki.duckyupdaterrework.core
 
 import dev.syoritohatsuki.duckyupdaterrework.core.api.ModrinthApi
 import dev.syoritohatsuki.duckyupdaterrework.core.api.models.Version
+import dev.syoritohatsuki.duckyupdaterrework.core.storage.Database
 import dev.syoritohatsuki.duckyupdaterrework.core.util.Hash
-import dev.syoritohatsuki.duckyupdaterrework.storage.Database
-import dev.syoritohatsuki.duckyupdaterrework.util.toInt
+import dev.syoritohatsuki.duckyupdaterrework.core.util.toInt
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
