+++ src/main/kotlin/dev/syoritohatsuki/duckyupdaterrework/core/util/Extensions.kt
@@ -1,0 +1,3 @@
+package dev.syoritohatsuki.duckyupdaterrework.core.util
+
+fun Boolean.toInt(): Int = if (this) 1 else 0
