+++ src/main/kotlin/dev/syoritohatsuki/duckyupdaterrework/core/util/Hash.kt
@@ -36,5 +36,4 @@ object Hash {
 
         else -> null
     }
-
 }
