--- src/main/kotlin/dev/syoritohatsuki/duckyupdaterrework/util/Extensions.kt
@@ -1,3 +1,0 @@
-package dev.syoritohatsuki.duckyupdaterrework.util
-
-fun Boolean.toInt(): Int = if (this) 1 else 0
